Overhaul how Graph status works
I realized, far too late, that there are TWO conditions that specify
whether the graph is "idle":

* The animations have all completed, nothing is moving
* The server has received all the data it needs from Spotify

Initially I was just using the former, which works as long as Spotify
takes <1000ms to respond. Today, Spotify is sluggish, and so the
loading indicator wasn't flashing while it waited for Spotify's data.

The solution is to keep track of two separate metrics, and to show
the loading symbol whenever either task is unfulfilled.